+++ src/quick-lint-js/i18n/translation-table-generated.h
@@ -19,7 +19,7 @@ using namespace std::literals::string_view_literals;
 
 constexpr std::uint32_t translation_table_locale_count = 5;
 constexpr std::uint16_t translation_table_mapping_table_size = 522;
-constexpr std::size_t translation_table_string_table_size = 79924;
+constexpr std::size_t translation_table_string_table_size = 79941;
 constexpr std::size_t translation_table_locale_table_size = 35;
 
 QLJS_CONSTEVAL std::uint16_t translation_table_const_look_up(
@@ -132,6 +132,7 @@ QLJS_CONSTEVAL std::uint16_t translation_table_const_look_up(
           "TypeScript 'declare function' is not allowed in JavaScript"sv,
           "TypeScript 'declare global' is not allowed in JavaScript"sv,
           "TypeScript 'declare global' is not allowed in namespaces"sv,
+          "TypeScript 'declare module' with string name is not allowed in namespaces"sv,
           "TypeScript 'declare {1}' is not allowed in JavaScript"sv,
           "TypeScript 'declare' fields are now allowed in JavaScript"sv,
           "TypeScript 'implements' is not allowed in JavaScript"sv,
@@ -411,7 +412,6 @@ QLJS_CONSTEVAL std::uint16_t translation_table_const_look_up(
           "missing value for object property"sv,
           "missing variable name"sv,
           "misspelled React attribute; write '{1}' instead"sv,
-          "module with string name is only allowed at the top level"sv,
           "move the 'extends' clause before 'implements' here"sv,
           "namespace starts here"sv,
           "new variable shadows existing variable"sv,